Fix test imports: move helper to tests/__init__.py
