+++ tests/test_multi_dataset.py
@@ -6,7 +6,7 @@ import hashlib
 import numpy as np
 import pytest
 
-from conftest import load_model_and_wait as _load_model_and_wait
+from tests import load_model_and_wait as _load_model_and_wait
 
 from vtsearch.utils.state_core import (
     DatasetContext,
